refactor(source/git): split git.go into auth.go + tls.go
Mirrors the existing test layout (auth_test.go, tls_test.go) which
already named what their production-side targets should be — but the
implementations lived in git.go alongside the Fetch entry point and
the cache/checkout helpers. Pure code reorg; behavior unchanged.

- auth.go: resolveAuth + isSSHURL + sshUserFromURL (covered by
  auth_test.go's existing 8 tests).
- tls.go: resolveTLS + httpsTransportMu (covered by tls_test.go's
  existing 6 tests).
- git.go: Fetcher type + Fetch + fetch + finalize + cache/checkout
  helpers (the actual Git operation surface).

git.go drops from 456 → 331 lines. Auth/TLS edits no longer pull
the full Fetch path into the editor scope.
